requirements.txt
gymnasium
stable-baselines3
numpy
pyyaml
matplotlib
